OneDrive/Desktop/Resume analyzer/resume-screening-ai/requirements.txt
# Resume Screening AI - Backend Requirements

# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23

# ML & NLP
sentence-transformers>=2.7.0
torch>=2.2.0
transformers>=4.35.2
scikit-learn>=1.3.2
numpy>=1.24.3

# Document Parsing (Advanced)
PyMuPDF==1.23.8  # Better PDF parsing than PyPDF2
python-docx==1.1.0
pdfminer.six==20221105  # Fallback PDF parser

# NLP & Text Processing
spacy>=3.7.0
nltk>=3.8.1
keybert>=0.8.4  # Keyword extraction using BERT

# Fuzzy Matching (already have rapidfuzz)
rapidfuzz>=3.14.1

# Web & HTTP
requests==2.31.0
aiofiles==23.2.1

# Data Processing
pandas==2.1.3

# Utilities
python-dotenv==1.0.0

# Logging
loguru==0.7.2

# Testing (optional but recommended)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
